backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.0.0
torchvision==0.15.0
pillow==10.0.0
pydantic==2.4.2
email-validator==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic-settings==2.0.3
